+++ packages/moonstone/VideoPlayer/MediaTitle.js
@@ -75,6 +75,7 @@ const MediaTitleBase = kind({
 			'infoComponents',
 			infoVisible ? 'visible' : 'hidden'
 		),
+		childrenRole: ({infoVisible}) => infoVisible ? 'alert' : '',
 		className: ({visible, styler}) => styler.append(
 			visible ? 'visible' : 'hidden'
 		),
@@ -84,7 +85,7 @@ const MediaTitleBase = kind({
 		})
 	},
 
-	render: ({children, childrenClassName, title, titleClassName, ...rest}) => {
+	render: ({childrenRole, children, childrenClassName, title, titleClassName, ...rest}) => {
 		delete rest.infoVisible;
 		delete rest.visible;
 
@@ -93,7 +94,7 @@ const MediaTitleBase = kind({
 				<MarqueeText className={titleClassName} marqueeOn="render">
 					{title}
 				</MarqueeText>
-				<div className={childrenClassName}>  {/* tabIndex={-1} */}
+				<div role={childrenRole} className={childrenClassName}>  {/* tabIndex={-1} */}
 					{children}
 				</div>
 			</div>