+++ packages/ui/ViewManager/ViewManager.js
@@ -31,6 +31,14 @@ class ViewManager extends React.Component {
 		arranger: shape,
 
 		/**
+		 * An object containing properties to be passed to each child.
+		 *
+		 * @type {Object}
+		 * @public
+		 */
+		childProps: PropTypes.object,
+
+		/**
 		 * Views to be managed. May be any renderable component including custom React components or
 		 * primitive DOM nodes.
 		 *
@@ -193,7 +201,7 @@ class ViewManager extends React.Component {
 	}
 
 	render () {
-		const {arranger, children, duration, end, index, noAnimation, start, enteringDelay, enteringProp, ...rest} = this.props;
+		const {arranger, childProps, children, duration, end, index, noAnimation, start, enteringDelay, enteringProp, ...rest} = this.props;
 		const {previousIndex, reverseTransition} = this;
 		const childrenList = React.Children.toArray(children);
 
@@ -210,7 +218,8 @@ class ViewManager extends React.Component {
 			previousIndex,
 			reverseTransition,
 			enteringDelay,
-			enteringProp
+			enteringProp,
+			childProps
 		});
 
 		delete rest.reverseTransition;